+++ src/components/Footer/Footer.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import Footer from './Footer';
+import Footer from '.';
 
 describe('Footer component', () => {
   it('should have the wizeline msg', () => {
