+++ src/components/Header/Header.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import Header from './Header';
+import Header from '.';
 
 describe('Header component', () => {
   it('should have the project title', () => {
