+++ src/components/Shimmer/index.jsx
@@ -1,0 +1,96 @@
+import { Placeholder, Row, Col } from 'react-bootstrap';
+
+const Shimmer = () => (
+  <Row data-testid='loader'>
+    <Col xs={6}>
+      <Placeholder animation='glow'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={1}></Col>
+    <Col xs={5}>
+      <Placeholder animation='wave'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={6}>
+      <Placeholder animation='glow'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={1}></Col>
+    <Col xs={5}>
+      <Placeholder animation='wave'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={6}>
+      <Placeholder animation='glow'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={1}></Col>
+    <Col xs={5}>
+      <Placeholder animation='wave'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={6}>
+      <Placeholder animation='glow'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={1}></Col>
+    <Col xs={5}>
+      <Placeholder animation='wave'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={6}>
+      <Placeholder animation='glow'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={1}></Col>
+    <Col xs={5}>
+      <Placeholder animation='wave'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={6}>
+      <Placeholder animation='glow'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={1}></Col>
+    <Col xs={5}>
+      <Placeholder animation='wave'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={6}>
+      <Placeholder animation='glow'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={1}></Col>
+    <Col xs={5}>
+      <Placeholder animation='wave'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={6}>
+      <Placeholder animation='glow'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+    <Col xs={1}></Col>
+    <Col xs={5}>
+      <Placeholder animation='wave'>
+        <Placeholder xs={12} size='lg' />
+      </Placeholder>
+    </Col>
+  </Row>
+);
+
+export default Shimmer;
