--- src/components/api/api.test.js
@@ -1,3 +1,0 @@
-describe('Api', () => {
-  it.todo('add all test for api');
-});
